# Speed up compilation by replacing `darling` with `attrs`