--- a/cached_proc_macro/Cargo.toml
+++ b/cached_proc_macro/Cargo.toml
@@ -18,7 +18,6 @@ proc-macro = true
 
 [dependencies]
 quote = "1.0.6"
-darling = "0.20.8"
 proc-macro2 = "1.0.49"
 syn = "2.0.52"
 strum-lite = "0.1.1"