--- a/cached_proc_macro/src/io_cached.rs
+++ b/cached_proc_macro/src/io_cached.rs
@@ -1,76 +1,70 @@
 use crate::helpers::*;
-use darling::ast::NestedMeta;
-use darling::FromMeta;
+use attrs::*;
 use proc_macro::TokenStream;
+use proc_macro2::Span;
 use quote::quote;
-use syn::spanned::Spanned;
+use syn::{parse::Parser as _, spanned::Spanned as _};
 use syn::{
     parse_macro_input, parse_str, Block, Expr, ExprClosure, GenericArgument, Ident, ItemFn,
     PathArguments, ReturnType, Signature, Type,
 };
 
-#[derive(FromMeta)]
-struct IOMacroArgs {
-    map_error: String,
-    #[darling(default)]
-    disk: bool,
-    #[darling(default)]
-    disk_dir: Option<String>,
-    #[darling(default)]
-    redis: bool,
-    #[darling(default)]
-    cache_prefix_block: Option<String>,
-    #[darling(default)]
-    name: Option<String>,
-    #[darling(default)]
-    time: Option<u64>,
-    #[darling(default)]
-    time_refresh: Option<bool>,
-    #[darling(default)]
-    key: Option<String>,
-    #[darling(default)]
-    convert: Option<String>,
-    #[darling(default)]
-    with_cached_flag: bool,
-    #[darling(default)]
-    ty: Option<String>,
-    #[darling(default)]
-    create: Option<String>,
-    #[darling(default)]
-    sync_to_disk_on_cache_change: Option<bool>,
-    #[darling(default)]
-    connection_config: Option<String>,
-}
-
 pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
-    let attr_args = match NestedMeta::parse_meta_list(args.into()) {
-        Ok(v) => v,
-        Err(e) => {
-            return TokenStream::from(darling::Error::from(e).write_errors());
-        }
+    let mut map_error = None::<String>;
+    let mut time = None::<u64>;
+
+    let mut time_refresh = None::<bool>;
+    let mut sync_to_disk_on_cache_change = None::<bool>;
+
+    let mut with_cached_flag = false;
+    let mut disk = false;
+    let mut redis = false;
+
+    let mut disk_dir = None::<String>;
+    let mut cache_prefix_block = None::<String>;
+    let mut name = None::<String>;
+    let mut key = None::<String>;
+    let mut convert = None::<String>;
+    let mut ty = None::<String>;
+    let mut create = None::<String>;
+    let mut connection_config = None::<String>;
+
+    match Attrs::new()
+        .once("map_error", with::eq(set::lit(&mut map_error)))
+        .once("time", with::eq(set::lit(&mut time)))
+        .once("time_refresh", with::eq(set::lit(&mut time_refresh)))
+        .once(
+            "sync_to_disk_on_cache_change",
+            with::eq(set::lit(&mut sync_to_disk_on_cache_change)),
+        )
+        .once("with_cached_flag", with::eq(on::lit(&mut with_cached_flag)))
+        .once("disk", with::eq(on::lit(&mut disk)))
+        .once("redis", with::eq(on::lit(&mut redis)))
+        .once("disk_dir", with::eq(set::lit(&mut disk_dir)))
+        .once(
+            "cache_prefix_block",
+            with::eq(set::lit(&mut cache_prefix_block)),
+        )
+        .once("name", with::eq(set::lit(&mut name)))
+        .once("key", with::eq(set::lit(&mut key)))
+        .once("convert", with::eq(set::lit(&mut convert)))
+        .once("ty", with::eq(set::lit(&mut ty)))
+        .once("create", with::eq(set::lit(&mut create)))
+        .once(
+            "connection_config",
+            with::eq(set::lit(&mut connection_config)),
+        )
+        .parse(args)
+    {
+        Ok(()) => {}
+        Err(e) => return e.into_compile_error().into(),
+    }
+    let Some(map_error) = map_error else {
+        return syn::Error::new(Span::call_site(), "Argument `map_error` is required")
+            .into_compile_error()
+            .into();
     };
-    let IOMacroArgs {
-        map_error,
-        disk,
-        disk_dir,
-        redis,
-        cache_prefix_block,
-        name,
-        time,
-        time_refresh,
-        key,
-        convert,
-        with_cached_flag,
-        ty,
-        create,
-        sync_to_disk_on_cache_change,
-        connection_config,
-    } = match IOMacroArgs::from_list(&attr_args) {
-        Ok(v) => v,
-        Err(e) => {
-            return TokenStream::from(e.write_errors());
-        }
-    };
+
     let ItemFn {
         attrs: mut attributes,
         vis: visibility,