About a 25% saving:

```console
$ hyperfine \
>    --warmup 1 \
>    --runs 10 \
>    --prepare 'git checkout attrs && cargo clean && rm Cargo.lock' \
>    --command-name attrs \
>    'cargo build --offline' \
>    --prepare 'git checkout master && cargo clean && rm Cargo.lock' \
>    --command-name master \
>    'cargo build --offline'
Benchmark 1: attrs
  Time (mean ± σ):      5.668 s ±  0.834 s    [User: 10.875 s, System: 2.352 s]
  Range (min … max):    4.863 s …  7.809 s    10 runs
 
Benchmark 2: master
  Time (mean ± σ):      7.090 s ±  0.174 s    [User: 12.745 s, System: 2.176 s]
  Range (min … max):    6.769 s …  7.323 s    10 runs
 
Summary
  attrs ran
    1.25 ± 0.19 times faster than master
```

Because we don't bottleneck compilation:
![wsl localhost_Ubuntu_home_aatif_code_cached_master html (1)](https://github.com/user-attachments/assets/d1d5903a-9fe3-4c41-899b-f070ed0435ce)

![wsl localhost_Ubuntu_home_aatif_code_cached_attrs html (1)](https://github.com/user-attachments/assets/01e269f9-1746-42aa-b516-666fa2bea0c1)

Note that I am the author of [`attrs`](https://docs.rs/attrs)