@jaemk 2025-05-01
hey @aatifsyed , you'll need to update the .github/workflows/build.yml like this https://github.com/jaemk/cached/pull/240/files#diff-5c3fa597431eda03ac3339ae6bf7f05e1a50d6fc7333679ec38e21b337cb6721L26-L38 to get the pipeline working again (unless that PR is merged before you see this)